#4360 Extract the NPC history mapping and share one observation repository interface
